pyproject.toml
[project]
name = "rag-dining"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "bitsandbytes==0.45.5",
    "chromadb==0.5.16",
    "einops==0.8.1",
    "faiss-cpu==1.8.0",
    "flagembedding==1.2.11",
    "onnxruntime==1.19.2",
    "optimum==1.23.3",
    "pandas==2.0.3",
    "sentence-transformers==3.2.1",
    "torch>=2.4.1",
    "transformers==4.46.3",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
